+++ assets/js/components/menus/index.js
@@ -1,4 +1,6 @@
 export { default as DefaultMenu } from './default_menu'
 export { default as LoginItem } from './login_item'
+export { default as MarkerControls } from './marker_controls'
+export { default as PlayerControls } from './player_controls'
 export { default as OverlayItem } from './overlay_item'
 export { default as SimMenu } from './sim_menu'