+++ assets/js/components/menus/sim_menu.js
@@ -4,9 +4,9 @@ import { NavLink } from 'react-router-dom'
 import { Icon, Menu } from 'semantic-ui-react'
 import { simOperations, simSelectors as Sim } from '../../modules/sim'
 import { SimTime } from '../../pages/sim'
-import { LoginItem, OverlayItem } from './'
+import { LoginItem, MarkerControls, PlayerControls } from './'
 
-const SimMenu = ({ changeMode, mode, play, pause, playing, stop }) => ([
+const SimMenu = ({ changeMode, mode }) => ([
   <Menu.Menu key={0}>
     <Menu.Item header exact as={NavLink} to='/videos' name='Exit Simulation' />
 
@@ -34,17 +34,8 @@ const SimMenu = ({ changeMode, mode, play, pause, playing, stop }) => ([
         />
       </a>
     </Menu.Item>
-
-    {playing
-      ? <Menu.Item icon='pause' onClick={() => pause()}/>
-      : <Menu.Item icon='play' onClick={() => play()}/>
-    }
-    <Menu.Item icon='stop' onClick={() => stop()} />
-
-    {mode == 'play' && <OverlayItem />}
-
-    {/* {user && <Menu.Item as={NavLink} to='/video/add' icon='plus' name='Add Video' />}
-    {user && <Menu.Item as={NavLink} to='/users' icon='users' name='Users' />} */}
+    {mode == 'play' && <PlayerControls />}
+    {mode == 'markers' && <MarkerControls />}
   </Menu.Menu>,
   <Menu.Menu key={1} position='right'>
     { mode != 'coords' &&
@@ -57,16 +48,6 @@ const SimMenu = ({ changeMode, mode, play, pause, playing, stop }) => ([
   </Menu.Menu>
 ])
 
-const mapStateToProps = store => ({
-  mode: Sim.mode(store),
-  playing: Sim.playing(store)
-})
-
-const mapDispatchToProps = {
-  changeMode: simOperations.changeMode,
-  play: simOperations.play,
-  pause: simOperations.pause,
-  stop: simOperations.stop
-}
-
+const mapStateToProps = store => ({ mode: Sim.mode(store) })
+const mapDispatchToProps = { changeMode: simOperations.changeMode }
 export default connect(mapStateToProps, mapDispatchToProps)(SimMenu)