+++ assets/js/modules/keys/index.js
@@ -1,4 +1,7 @@
 import reducer from "./reducers"
 
+export { default as keySelectors } from "./selectors"
 export { default as keyOperations } from "./operations"
-export default reducer;
+export { default as keyTypes } from "./types"
+
+export default reducer