+++ assets/js/modules/keys/selectors.js
@@ -1,0 +1,5 @@
+const pressed = state => state.keys
+
+export default {
+  pressed
+}