+++ assets/js/modules/sim/types.js
@@ -1,43 +1,27 @@
-const CHANGE_MODE = 'sim/CHANGE_MODE'
-const CHANGE_PLAYER_STATE = 'sim/CHANGE_PLAYER_STATE'
-const CLEAR_ERROR = 'sim/CLEAR_ERROR'
-const ERROR = 'sim/ERROR'
-const JOIN = "sim/JOIN"
-const JUMP = "sim/JUMP"
-const LEAVE = "sim/LEAVE"
-const LOAD_PLAYER = 'sim/LOAD_PLAYER'
-const PLAY = 'sim/PLAY'
-const PAUSE = 'sim/PAUSE'
-const STOP = 'sim/STOP'
-const RESIZE = "sim/RESIZE"
-const SELECT_AGENT = "sim/SELECT_AGENT"
-const SET_OVERLAY = 'sim/SET_OVERLAY'
-const TICK = "sim/TICK"
-const UPDATE = "sim/UPDATE"
-const UPDATE_VIDEO = "sim/UPDATE_VIDEO"
-const VIDEO_NOT_FOUND = "sim/VIDEO_NOT_FOUND"
-
 export default {
   // overlays
   ADD_OVERLAY: 'sim/ADD_OVERLAY',
   REMOVE_OVERLAY: 'sim/REMOVE_OVERLAY',
+  SET_OVERLAY: 'sim/SET_OVERLAY',
 
-  CHANGE_MODE,
-  CHANGE_PLAYER_STATE,
-  CLEAR_ERROR,
-  ERROR,
-  JOIN,
-  JUMP,
-  LEAVE,
-  LOAD_PLAYER,
-  PLAY,
-  PAUSE,
-  STOP,
-  RESIZE,
-  SELECT_AGENT,
-  SET_OVERLAY,
-  TICK,
-  UPDATE,
-  UPDATE_VIDEO,
-  VIDEO_NOT_FOUND
-};
+  // player
+  LOAD_PLAYER: 'sim/LOAD_PLAYER',
+  CHANGE_PLAYER_STATE: 'sim/CHANGE_PLAYER_STATE',
+  PLAY: 'sim/PLAY',
+  PAUSE: 'sim/PAUSE',
+  STOP: 'sim/STOP',
+  TICK: 'sim/TICK',
+
+  CHANGE_JUMP_INTERVAL: 'sim/CHANGE_JUMP_INTERVAL',
+  CHANGE_MODE: 'sim/CHANGE_MODE',
+  CLEAR_ERROR: 'sim/CLEAR_ERROR',
+  ERROR: 'sim/ERROR',
+  JOIN: 'sim/JOIN',
+  JUMP: 'sim/JUMP',
+  LEAVE: 'sim/LEAVE',
+  RESIZE: 'sim/RESIZE',
+  SELECT_AGENT: 'sim/SELECT_AGENT',
+  UPDATE: 'sim/UPDATE',
+  UPDATE_VIDEO: 'sim/UPDATE_VIDEO',
+  VIDEO_NOT_FOUND: 'sim/VIDEO_NOT_FOUND'
+}