+++ assets/js/pages/sim/markers.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { simOperations, simSelectors as Sim } from '../../modules/sim'
 
-const Markers = ({ agents, agentSelected, comparisonMode, overlay, update }) => (
+const Markers = ({ agents, agentSelected, comparisonMode, mode, overlay, update }) => (
   Object.keys(agents).map(( i ) => {
     const id = parseInt(i)
     let className = 'marker'
@@ -13,8 +13,8 @@ const Markers = ({ agents, agentSelected, comparisonMode, overlay, update }) => 
       <div
         key={id}
         className={ className }
-        onMouseEnter={() => update({ agentHovered: id })}
-        onMouseLeave={() => update({ agentHovered: null })}
+        onMouseEnter={() => mode == 'markers' ? update({ agentHovered: id }) : null }
+        onMouseLeave={() => mode == 'markers' ? update({ agentHovered: null }) : null }
         style={{
           left: 100 * agents[i].x+"%",
           top: 100 * agents[i].y+"%"
@@ -33,6 +33,7 @@ const mapStateToProps = ( state, ownProps ) => ({
         : Sim.getRelPositions(state)
       ),
   agentSelected: state.sim.agentSelected,
+  mode: Sim.mode(state),
   overlay: Sim.overlay(state)
 })